refactor(app): extract useTabs hook (full tab lifecycle)
Phase 1 step 6 of the god-file cleanup plan. Move tab create / switch /
update / close / undo-close into src/hooks/useTabs.ts. The hook owns:

- TAB_MIRROR_KEYS + TERMINAL_REPLAY_MAX exports.
- pendingTabOpens, closedTabsRef, autoRestoredSessionIdsRef refs.
- updateTab, updateActiveTab, applyShareModeToTab.
- dispatchTerminalReplay, setActiveTab, setActiveSubTab.
- newTab, newShellTab, autoRestoreDiscoveredSessions.
- pushClosedTab, reopenLastClosedTab, closeTab, closeTabNow.
- The /shellTabActive + /agentTabActive mirror effect (kind-derived
  layout-visible flags).

App passes context callbacks (recomputeModelPicker stays local since
the sidebar shape is broader than tabs; clearActiveProject /
setActiveProjectById from the still-inline project orchestrators;
shell config refs from the boot config effect / settings panel apply
path).

App.tsx 5568 -> 4978 lines (-590; total -1246 from the original 6224,
-20%). tsc clean, eslint clean, 305/305 tests pass. No behavior
change (the original closeTabNow had two branches that fired
identical tab_close IPC; the hook collapses them).
